refactor: rename gtk-dom-event-bridge to event-bridge, integrate into widgets
Rename @gjsify/gtk-dom-event-bridge → @gjsify/event-bridge for brevity.
Integrate attachEventControllers() into both Canvas2DWidget and
CanvasWebGLWidget so GTK events (mouse, keyboard) are bridged to DOM
events on the canvas element. Confetti example gains mouse-driven wind.

Also split CLAUDE.md into AGENTS.md (project instructions) and
CLAUDE.md (compact pointer), per user restructuring.
