+++ packages/dom/canvas2d/package.json
@@ -38,7 +38,7 @@
         "@girs/pango-1.0": "^1.57.0-4.0.0-beta.42",
         "@girs/pangocairo-1.0": "^1.0.0-4.0.0-beta.42",
         "@gjsify/dom-elements": "workspace:^",
-        "@gjsify/gtk-dom-event-bridge": "workspace:^",
+        "@gjsify/event-bridge": "workspace:^",
         "@gjsify/utils": "workspace:^"
     },
     "devDependencies": {
