+++ packages/dom/canvas2d/src/canvas-drawing-area.ts
@@ -6,7 +6,7 @@ import GObject from 'gi://GObject';
 import GLib from 'gi://GLib?version=2.0';
 import Gtk from 'gi://Gtk?version=4.0';
 import { HTMLCanvasElement as GjsifyHTMLCanvasElement } from '@gjsify/dom-elements';
-import { attachEventControllers } from '@gjsify/gtk-dom-event-bridge';
+import { attachEventControllers } from '@gjsify/event-bridge';
 import { CanvasRenderingContext2D } from './canvas-rendering-context-2d.js';
 
 type Canvas2DReadyCallback = (canvas: globalThis.HTMLCanvasElement, ctx: CanvasRenderingContext2D) => void;
