+++ packages/dom/event-bridge/src/index.ts
