+++ packages/dom/event-bridge/tsconfig.json
