+++ packages/dom/webgl/package.json
@@ -48,7 +48,7 @@
         "@girs/gtk-4.0": "4.22.1-4.0.0-beta.42",
         "@girs/gwebgl-0.1": "workspace:^",
         "@gjsify/dom-elements": "workspace:^",
-        "@gjsify/gtk-dom-event-bridge": "workspace:^",
+        "@gjsify/event-bridge": "workspace:^",
         "@gjsify/utils": "workspace:^",
         "@types/bit-twiddle": "^1.0.3",
         "bit-twiddle": "^1.0.2",
