+++ packages/dom/webgl/src/ts/canvas-webgl-widget.ts
@@ -6,7 +6,7 @@ import GLib from 'gi://GLib?version=2.0';
 import Gtk from 'gi://Gtk?version=4.0';
 import { HTMLCanvasElement as OurHTMLCanvasElement } from './html-canvas-element.js';
 import type { WebGLRenderingContext as OurWebGLRenderingContext } from './webgl-rendering-context.js';
-import { attachEventControllers } from '@gjsify/gtk-dom-event-bridge';
+import { attachEventControllers } from '@gjsify/event-bridge';
 
 // Public callback type uses globalThis.HTMLCanvasElement (lib.dom) so callers can pass the
 // canvas directly to WebGL demos that type their canvas parameter as HTMLCanvasElement.
